move PrefixRecord to records.py
